Constrain dashboard queries to selected period
If/when the SolidCacheDashboard::CacheEvent table grows to a certain
size, performance quickly plummets.

This attempts to ensure that the dashboard loads only the relevant
records to avoid this.

Note that the counts are also constrained to the selected period - this
seems appropriate, but perhaps the more complete counts are actually
intended.