Restore recent entries/events, fix bugs
This corrects a bug that raised an error on a blank chart_period, which
was exposed by a possible bug in the navbar in which the hidden field
has no value.

This also restores the recent events and recent entries to show the most
recent 5 - instead of pulling 5 from a query of all rows, they are
pulled from a separate query and ordered and limited.